+++ src/shared/components/NavBar/NavBar.styles.tsx
