+++ src/shared/components/NavBar/NavBar.tsx
@@ -1,5 +1,5 @@
 import { Typography } from '@mui/material';
-import { StyledNavBar } from './navBar.styles';
+import { StyledNavBar } from './NavBar.styles';
 import MenuIcon from '@mui/icons-material/Menu';
 import { FC } from 'react';
 
@@ -11,7 +11,7 @@ interface NavBarProps {
 const NavBar: FC<NavBarProps> = ({ smallScreen, toggle }) => {
     return (
         <StyledNavBar>
-            {smallScreen ? <MenuIcon onClick={toggle} /> : null}
+            {smallScreen && <MenuIcon onClick={toggle} />}
             <Typography>NavBar</Typography>
         </StyledNavBar>
     );
